Move TD46 stuff to own file
