+++ src/openrct2/rct12/RCT12.h
@@ -377,19 +377,9 @@ constexpr uint8_t kTD46GForcesMultiplier = 32;
 constexpr uint8_t kRCT12InversionAndHoleMask = 0b00011111;
 constexpr uint8_t kRCT12RideNumDropsMask = 0b00111111;
 
-struct TrackDesign;
-struct TrackDesignTrackElement;
 enum class RideColourScheme : uint8_t;
 enum class BannerFlag : uint8_t;
 using BannerFlags = FlagHolder<uint8_t, BannerFlag>;
-
-enum class RCT12TrackDesignVersion : uint8_t
-{
-    TD4,
-    TD4_AA,
-    TD6,
-    unknown
-};
 
 enum
 {
@@ -508,46 +498,6 @@ struct RCT12xy8
     }
 };
 static_assert(sizeof(RCT12xy8) == 2);
-
-enum class TD46MazeElementType : uint8_t
-{
-    Entrance = (1 << 3),
-    Exit = (1 << 7)
-};
-
-/* Maze Element entry   size: 0x04 */
-struct TD46MazeElement
-{
-    union
-    {
-        uint32_t All;
-        struct
-        {
-            int8_t x;
-            int8_t y;
-            union
-            {
-                uint16_t MazeEntry;
-                struct
-                {
-                    uint8_t Direction;
-                    uint8_t Type;
-                };
-            };
-        };
-    };
-
-    constexpr bool IsEntrance() const
-    {
-        return Type == EnumValue(TD46MazeElementType::Entrance);
-    }
-
-    constexpr bool IsExit() const
-    {
-        return Type == EnumValue(TD46MazeElementType::Exit);
-    }
-};
-static_assert(sizeof(TD46MazeElement) == 0x04);
 
 /* Track Element entry  size: 0x02 */
 struct TD46TrackElement
@@ -1293,19 +1243,6 @@ std::vector<RideId> RCT12GetRidesBeenOn(T* srcPeep)
     }
     return ridesBeenOn;
 }
-
-enum class TD46Flags : uint8_t
-{
-    StationId = 0b00000011,
-    SpeedOrSeatRotation = 0b00001111,
-    ColourScheme = 0b00110000,
-    IsInverted = 0b01000000,
-    HasChain = 0b10000000,
-};
-
-void ConvertFromTD46Flags(TrackDesignTrackElement& target, uint8_t flags);
-uint8_t ConvertToTD46Flags(const TrackDesignTrackElement& source);
-void ImportMazeElement(TrackDesign& td, const TD46MazeElement& td46MazeElement);
 
 namespace OpenRCT2::RCT12
 {
